1.1.1+2025/06/24
Se agrega dependencia para documentar

+++ src/main/java/com/SAFE_Rescue/API_Administrador/repository/BomberoRepository.java
@@ -12,8 +12,8 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface BomberoRepository extends JpaRepository<Bombero , Long> {
 
-    public boolean existsByRun(Long run);
+    public boolean existsByRun(Integer run);
 
-    public boolean existsByTelefono(Long telefono);
+    public boolean existsByTelefono(Integer telefono);
 
 }
